pyproject.toml
[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[project]
name = "robot_framework"
version = "1.0.0"
authors = [
  { name="ITK Development", email="itk-rpa@mkb.aarhus.dk" },
]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
]
dependencies = [
    "OpenOrchestrator == 2.*",
    "Pillow == 10.*",
    "itk-dev-shared-components == 2.*",
    "pyyaml"
]

[project.optional-dependencies]
dev = [
  "pylint",
  "flake8"
]
